Crate second_stack

Structs

Stack

A Stack that is managed separately from the threadlocal one. Typically, using the threadlocal APIs is encouraged because they enable sharing across libraries, where each re-use lowers the amortized cost of maintaining allocations. But, if full control is necessary this API may be used.

Functions

buffer

Buffers an iterator to a slice on the threadlocal stack and gives temporary access to that slice. Panics when running out of memory if the iterator is unbounded.

uninit

Place a potentially very large value on the threadlocal second stack.

uninit_slice

Allocates an uninit slice from the threadlocal stack.