second_stack

Struct Stack

pub struct Stack(/* private fields */);

A Stack that is managed separately from the threadlocal one. Typically, using the threadlocal APIs is encouraged because they enable sharing across libraries, where each re-use lowers the amortized cost of maintaining allocations. But, if full control is necessary this API may be used.

Implementations

impl Stack

pub fn new() -> Self

pub fn uninit<T, R, F>(&self, f: F) -> R
where F: FnOnce(&mut MaybeUninit<T>) -> R,

Place a potentially very large value on this stack.

pub fn uninit_slice<T, F, R>(&self, len: usize, f: F) -> R
where F: FnOnce(&mut [MaybeUninit<T>]) -> R,

Allocates an uninit slice from this stack.

pub fn buffer<T, F, R, I>(&self, i: I, f: F) -> R
where I: Iterator<Item = T>, F: FnOnce(&mut [T]) -> R,

Buffers an iterator to a slice on this stack and gives temporary access to that slice. Do not use with an unbounded iterator, because this will eventually run out of memory and panic.